+++ app/cmd/splat3/recruit/regular_recruit.js
@@ -8,7 +8,8 @@ const { checkFes, getRegularData, fetchSchedule } = require('../../../common/api
 const { isNotEmpty, isEmpty, sleep } = require('../../../common');
 const { searchChannelIdByName } = require('../../../manager/channelManager');
 const { createRoundRect, drawArcImage, fillTextWithStroke } = require('../../../common/canvas_components');
-const { recruitActionRow, setButtonDisable, recruitDeleteButton, unlockChannelButton } = require('../../../common/button_components');
+const { recruitActionRow, recruitDeleteButton, unlockChannelButton } = require('../../../buttons/create_recruit_buttons');
+const { setButtonDisable } = require('../../../common/button_components');
 const { AttachmentBuilder, ChannelType, PermissionsBitField } = require('discord.js');
 const { dateformat, formatDatetime } = require('../../../common/convert_datetime');
